--- dynamic_portfolio_results/.ipynb_checkpoints/models-checkpoint.py
@@ -1,31 +1,0 @@
-from otree.api import (
-    models, widgets, BaseConstants, BaseSubsession, BaseGroup, BasePlayer,
-    Currency as c, currency_range
-)
-import numpy
-import json
-
-
-author = 'Your name here'
-
-doc = """
-Your app description
-"""
-
-
-class Constants(BaseConstants):
-    name_in_url = 'dynamic_portfolio_results'
-    players_per_group = None
-    num_rounds = 2
-
-
-class Subsession(BaseSubsession):
-    pass
-
-
-class Group(BaseGroup):
-    pass
-
-
-class Player(BasePlayer):
-    pass
