+++ settings.py
@@ -20,7 +20,7 @@ APP_KITS = {
             'task1_prac_tr1',
             'task1_quiz_tr12',
             'task1_tr1',
-            'dynamic_portfolio_results',
+            'task1_results_tr12',
             'new_task2_instructions',
             'new_task2_practice',
             'new_task2_quiz',
@@ -43,7 +43,7 @@ APP_KITS = {
             'task1_prac_tr2',
             'task1_quiz_tr12',
             'task1_tr2',
-            'dynamic_portfolio_results',
+            'task1_results_tr12',
             'new_task2_instructions',
             'new_task2_practice_chp',
             'new_task2_quiz',
@@ -98,7 +98,7 @@ SESSION_CONFIGS = [
                       'task1_tr1',
                       'task1_tr2',
                       'tr3_dynamic_pooled',
-                      'dynamic_portfolio_results',
+                      'task1_results_tr12',
                       'tr3_dynamic_pooled_results',
                       'new_task2_instructions',
                       'new_task2_practice',
